fix: do not render multi-incoming vertices node to avoid tree bloating 💥
This moves tree components in a Readable store, instead of a 'root node' (App.svelte) variable.
It avoids, when tree has to be updated (CRUD of nodes) to pass messages
from deep components up to the root node. A store is used instead, and
lists of multi-incoming vertices nodes are dynamically computed.

This changes a bit how users will use the tree, as the multi-incoming vertices nodes' children
aren't rendered anymore. They have to be selected from the appropriate list.